+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix(`AbstractState`): Make `$container` protected, so it can be extended by
   child classes
 * build: Create `ApiMapperInterface` and use it for the service ID.
+* refactor: Move `mapCollection()` to separate interface.
 
 ## 0.5.0